Use PathBuf for all field in Repo struct
This commit switches the Repo stuct to use PathBufs for
all of its fields. Because PathBufs are heap allocated
(similar to Strings) this removes the need to have a boxed
pointer for all the Repo fields which was required when
using the Path type. Boxing Path was required because the
size of the Path cannot be known at compile time.