Move repo setup test and helper into src/
This commit moves the helper setup test into the object.rs file.
The happy path when adding unit tests is to create a module that
lives in the same file as the code you intend to test. Prior to
this commit this test was located in the tests/object_tests file
any tests located in tests/ should be integration level tests an
check things from an outside prespective.