Use PathBuf for path type in Config struct
This commit changes the type for for the path field in the Config
struct. It also adds the repo_path as an arg to the Config::new fn,
this change will help with testing and is needed for the git init
command. An override for the hard codeded "." path will be added in
a future commit.