Add project level Error enum
This commit adds a project level Error enum to capture all the
possible errors which may occur in the project. Prior to this
commit strings were returned on the Error side of the Result type.
The thiserror crate is used to provide human readable error messages
when an error enum instance needs to be printed.