Read git config during Repo creation
This commit adds a std::convert::From impl for std::io::Error
to the Project's Error type. This allows for the use of the ?
operator when calling functions which return std::io::Error.
This commit reads the contents of the git config file using
fs::read_to_string and stores this information in the Repo
type under the gitconf field.