Move git repo check into Repo struct creation
This commit moves the check in utils::git_repo_or_err into the
creation of the Repo struct. Moving the check into Repo accounts
for commands like 'init' which create a git repo. Expecting an
existing git repo in the Config construction is incorrect because
not all commands require that they're run from an existing git repo.