Add helper which finds a .git dir and then creates Repo inst
This commit adds a helper function which searches for a .git dir
on the path given in the command Config by looking at all the
parent dirs until root. If a .git dir is found the path at that
location is used in the Repo construction, if no .git dir is found
a Error::NotAGitRepo is returned.