Add write_object function
This commit adds a function which takes a file path, adds the appropriate
headers to the file contents, and then hashes those bytes, returning the
hash. If a Repo argument is supplied the function takes the extra step of
compressing the information decribed above and writing that information at
a path in .git/objects determined by the hash.