Convert Nom errors to proj errors
This commit changes the return type of the top level parse_git_obj
function so that it returns an instance of NomError from the project's
Error enum. This change makes the parse_git_obj func's return type
compatible with the error type used throughout the project which
allows use of the ? operator.