Expose cat-file command via run_cmd
This commit wires up the cat-file command so that it can be run from
the rusty-git command line tool. Rusty-git's version of cat-file doesn't
expect to be given the obj type like git's,
e.g., 'git cat-file <obj type> <sha>'

This commit also adds a flag to the run_cmd function. This flag controls
whether a Repo will be added as an arg to the command being run. The
presence of a Repo struct is used to control the behavior of the
hash-object command. The flag is hard coded to false when run from the
command line to prevent rusty-git from affecting an existing git repo.
In tests which take place in a temp dir this flag is set to true so
that the round trip behavior of placing objects in .git/objects can be
tested.