Change the contents field to owned type in GitObject
This commit changes the contents field in GitObject to a Vec<u8>.
Prior to this commit the contents field was a &[u8]. Having a ref
to the bytes Nom parses was causing issues when reading a file into
a variable in a func and then trying to return a struct with a ref
to that local variable. This changes converts the bytes Nom parses
to an owned Vec in the parsing function. This change also breaks
apart the funcs responsible for parsing of a git object to a string
vs. parsing it into a GitObject.