Add sha to KvsMsg struct
This commit adds a sha field to the KvsMsg struct that's used when
parsing the key-value pairs and the message found in commits and tags.
The sha is needed after parsing but isn't found in the contents of
commits and tags but it is known at parse time. Including the sha
will be useful later on when we need to display or use the commit sha.