Add Tree struct and parser for Git trees
This commit adjusts the parse_git_tree_leaf function so that it returns a
Nom IResult which allows it to be used with many1(). The elements in the
triple that this function returns are later converted into owned strings
in the parse_git_tree func which returns an instance of the new Tree type.
These parsers will be used in future commits when Git tree objects are
consumed.