Change how shas in git tree files are parsed
Prior to this commit shas in git tree files were treated as utf-8
during parsing which was incorrect. Git stores sha-1 hashes using
20 bytes where each byte stores two hexidecimal characters to save
space. This needs to be taken into account when parsing the shas
in git tree files. This commit uses the Rust formater to convert
each byte in a sha into a hexidecimal pair.