Add checkout command
This commit adds a restricted version of the git checkout command. When
running the command a user must first create an empty directory and pass
the path to that directory in as an argument. The  checkout command will
populate the given directory with the contents of the worktree at the
given commit.