Fix bug when stripping prefix from ref path
Prior to this commit the prefix being stripped from a ref's path was
hardcoded to "./.git". This works for the general case when the
command is run in a git repo but not during testing when tempdirs are
used. Now repo.gitdir is used to figure out which prefix to strip and
accounts for the tempdirs used during testing.