Add lightweight git tag command
This commit implements the lightweight version of the git tag command.
When a user runs a 'rusty-git tag foo <object-sha>' command, a file
named foo containing the object-sha (default: HEAD) will be created in
the .git/refs/tags/ directory. The heavyweight version of this command
which creates a tag object will be implemented in a later commit.