Change Nom error conversion to generic type
This commit changes how Nom errors are converted to rusty-git errors
using a generic type T that must implement the Debug interface. This
change will be used in following commits when bit parsers are needed
and the type signature of the errors they return don't match the &[u8]
of the byte parsers.