Adjust parse_git_index_entry return type
This commit changes the return type of the parse_git_index_entry
func so that it returns a Nom IResult. This allows the parser to
be used with other Nom parser combinators. The trailing padding
bytes in index entries are now taken into account and consumed
by the parser. The test has been updated to reflect this change
and the unneeded bytes which represented the whole git index file
in the test have been removed.