Add utils func that returns file sha pairs from git Tree object
This commit adds a utils func that knows how to drill down into
nested git Tree objects and returns a set of (file path, sha)
pairs. This format matches what is saved in the git Index and
can be used to compare which files are staged currently vs. the
state of the worktree from the HEAD commit.