Add helper that returns diff between index and HEAD commit tree
This commit adds a utils helper func that returns the file name and
sha for any files that are different in the index vs. what was in the
HEAD commit tree. This function will be used when implementing the git
status command to show which files are staged to be commited when running
git status.