Add helper that checks for .rusty-git-allowed file
This commit adds a helper function which errors if a .rusty-git-allowed
file doesn't exist in the worktree. This check will be used in later
commits to prevent rusty-git from performing any operations that alter
the git index or write objects unless the user opts in and allows these
changes by creating a .rusty-git-allowed file.