Impl ToBinary trait for Index
This commit implements the to_binary function for the Index
struct. This function converts the in men Index struct into
bytes that match the format of the .git/index file. The Index
struct now has an extensions field that captures the extension
information in .git/index file, this addition was needed to be
able to round trip the contents of the index file.