Change return type from write_object
This commit changes the return type from the write_object function.
Prior to this commit the String representation of the sha1 for an
object was returned. Now a sha1::Digest of the hash is returned so
the caller can decide between the byte or String representation of
the hash. This will be used in subsequent commits when hashing an
object that needs to be added to the git index.