Add helper which adds an IndexEntry to Index.entries
This commit adds a helper func that can add an IndexEntry to the
sorted Index.entries field. The Ord trait was implemented for
IndexEntry to allow for the vector to support binary_search. The
name of each IndexEntry is used in the ordering of Index.entries.
If an existing entry is found in the Index it is replaced by the
new entry. This helper will be used when implementing the git add
command.