Fix bug when calling rusty-git init
This commit fixes a bug where the git_repo_or_err function was being
called when a new instance of the internal struct Repo is being
created. For most commands this is fine and the behavior that you'd
want but it causes init to fail because the directory is not yet a
git repo. This commit changes this behavior so the Repo stuct isn't
created during the init command.