Add check for the existence of .git/index
This commit adds a check for the existence of the .git/index file before
the status command is run. If the .git/index file doesn't exist it
indicates that a user hasn't told git to track any files yet. Git handles
this situation differently and displays a partial status, this tool prompts
the user to use the add command to start tracking a file which will trigger
the index creation.