Fix bug when status is run before any commits exist
This change fixes a bug that could happen when running the status
command when no commit objects exist yet. When commits exist you
want to compare the names and shas of the files in the repo to the
names and shas of the files referenced in the HEAD commit. If there
are no commits there is nothing to compare against so an empty HashSet
is now used instead of returning an error.