Adjust error handling in IndexEntry parser
This commit adjusts the type of Nom error that's returned when parsing
fails in the parse_git_index_entry function. Prior to this commit a
Nom::Failure was returned which causes the parsing to pop and bubble
the error up to the user. This isn't the behavior needed when parsing
an entry. The parser might be called on the extension information or
sha that follows the last entry in the index file, or there may be no
entries in the index at all. In this case the parser should return a
Nom::Error with ErrorKind::Many0, this combination allows the parser
to continue parsing and tries the next parser with the input that was
passed to the parse_git_index_entry function.