Store 20 byte sha in TreeLeaf
This commit changes the data type of the sha field in the TreeLeaf
struct used to represent git tree entries from a String to Vec<u8>.
By keeping the 20 byte representation of the sha unchanged converting
a TreeLeaf back into bytes is easier. Now the conversion from the 20
byte sha to a 40 character string is done only when the sha needs to
be displayed.