Change type returned from read_object
This commit changes the type returned from the read_object function. Prior
to this commit the intermediate type objects::GitObject was returned and
then the contents inside of this type were later parsed depending on whether
it was a blob, commit, or tree. Now the parsing is done inside read_object
and an object_mods::GitObj::{Blob, Commit, Tree} is returned directly.

This change helps to clean up both the callers and down stream functions
which can now all expect the same GitObj type. Each sub type in GitObj
implements the AsBytes interface which will be useful when these objects
need to be written during the 'git commit' command.