Change in mem representation for Commit objects
This commit changes the struct used to represent git commit objects
while the program is running. Prior to this commit the raw bytes were
stored in a hashmap that could be used to look up information about
the commit. This was done as a bit of a hack to make round tripping
the commit bytes easier. Now the commit is stored in a struct with
all the info in a commit having named field and String value.
e.g. commit.tree <String sha>. This change will make the 'git commit'
command easier to implement because creating Commit instances is
simplified.