Only commit if there are staged files in the index
Prior to this commit the 'rusty-git commit -m <msg>' command would
create a git commit object everytime it was run. This would happen
even if there were no changed files staged in the index. This commit
adds a check and returns early from the commit command if there are
no changes to be committed staged in the index.