Remove unneeded returns
This commit removes the unneeded returns found throughout the project
pointed out by clippy. Because Rust is an expression oriented language
the last expression in any function will be evaluated and its result
is returned implicitly. Return is only needed when you want to indicate
an early return from an expression. Including an ending ; indicates that
something is a statement and omitting the ; means that something is an
expression and returns its result.