Change calls where an unnecessary reference was being passed
Clippy pointed out a few locations where functions were being called with
a reference to an already existing reference. The added & is unnecessary
as the complier was immediatley dereferencing these already. Removing the
extra layer of borrowing is more clear because using &foo suggests that
the receiver of &foo is borrowing foo.