Remove extra layer of unwrapping/wrapping with Ok(foo?)
This commit changes the expression returned by two functions to be the
return value of the helper functions they call. In this case because
both helpers have the same return type as the calling functions the value
of the helpers can be returned directly without the need to unwrap the
helper's result with the '?' operator and then wrap that result back in
an Ok(...).