Change write to write_all
This commit changes write to write_all, this is needed because write is
not guaranteed to process the entire buffer. In this case I don't need
to worry about partial writes so using write_all is preferable and avoids
the need to check that all the intented bytes were written.