Remove redundant clone calls
This commit removes redundant clone and to_owned calls which can result in
unnecessary allocations if the compiler is unable to recognize and eliminate
these useless allocations. Redundant clones happen when a value is going to be
dropped without further use but a call to .clone() or one of its variants is
made to clone the value.