Replace format! with writeln!
Using format! allocates another String on the heap which isn't necessary. In
this case you can write directly to the string which avoids this second
allocation. The Result from write needs to be handled in order to use the ?
operator but it should never fail when used in this way.