Put GitObj::Commit in a Box
This commit uses a Box to store the Commit variant of the GitObj enum. This is
preferable because enum size is bounded by the size of the largest variant. In
this case the Commit variant was 240 bytes while the next largest variant was 32
bytes. By using Box the enum stores a pointer to the Box and the actual data
inside the Box is stored on the heap. Doing this reduces the size of the enum
and doesn't waste space on the stack when using one of the variants that isn't
Commit.